refactor : HttpMethod 생성시 예외 처리
