+++ app/src/main/java/com/codesoom/http/HttpMethod.java
@@ -1,4 +1,6 @@
-package com.codesoom.controller;
+package com.codesoom.http;
+
+import com.codesoom.exception.MethodNotExistException;
 
 public enum HttpMethod {
     GET, POST, PUT, DELETE;
@@ -6,13 +8,24 @@ public enum HttpMethod {
     public boolean isPost() {
         return this == POST;
     }
+
     public boolean isGet() {
         return this == GET;
     }
+
     public boolean isPut() {
         return this == PUT;
     }
+
     public boolean isDelete() {
         return this == DELETE;
+    }
+
+    public static HttpMethod of(String method) {
+        try {
+            return HttpMethod.valueOf(method);
+        } catch (IllegalArgumentException e) {
+            throw new MethodNotExistException();
+        }
     }
 }
